+++ demo/src/test/java/com/example/test/pages/KBBMOOCProject/CarForSalePage.java
@@ -34,23 +34,4 @@ public class CarForSalePage extends KbbBasePage{
         return true;
 
     }
-
-    public int getMiles()
-    {
-        WebElement miles = keyword.findWebElementByXpath("//span[@class='text-bold'][1]");
-        String strMiles = miles.getText().replace(" miles", "");
-        double dbMiles = Double.parseDouble(strMiles.replaceAll(",", ""));
-        int intMiles = (int) dbMiles;
-        return intMiles;
-    }
-
-    public boolean comparasionMilesBetweenCheckboxAndResult() throws InterruptedException
-    {
-        keyword.waitForPageToLoad();
-        if(getMiles()<15000)
-        {
-            return true;
-        }
-        return false;
-    }
 }
